requirements.txt
fastapi>=0.103.2
uvicorn>=0.23.2
starlette>=0.27.0
mysql-connector-python>=8.1.0
openai>=1.3.3
auth0-python>=4.7.1